added fullDateRangeImport function
This new function will query the api multiple times in order to get the entire date entered. This can easily be used to ingest the entire database if needed.

NOTE the test code that is in the init main function generates a 1GB text file at the end. This is the result of 1 years worth of CVE's imported